utils: Move treeadapter et al to Orange/utils
