+++ Orange/utils/tree/treeadapter.py
