+++ i18n/si/msgs.jaml
@@ -4217,6 +4217,59 @@ tests/sql/test_sql_table.py:
             postgres: false
             mssql: false
     __main__: false
+utils/tree/rules.py:
+    class `DiscreteRule`:
+        def `merge_with`:
+            Merged two discrete rules `%s` and `%s`: false
+        def `description`:
+            {} {}: false
+            =: false
+            ≠: false
+        def `__str__`:
+            {} {} {}: false
+            =: false
+            ≠: false
+        def `__repr__`:
+            DiscreteRule(attr_name='%s', equals=%s, value=%s): false
+    class `ContinuousRule`:
+        def `merge_with`:
+            'Continuous rules can currently only be ': false
+            merged with other continuous rules: false
+        def `description`:
+            %s %.3f: false
+            >: false
+            ≤: false
+        def `__str__`:
+            %s %s %.3f: false
+            >: false
+            ≤: false
+        def `__repr__`:
+            "ContinuousRule(attr_name='%s', greater=%s, value=%s, ": false
+            inclusive=%s): false
+    class `IntervalRule`:
+        def `__init__`:
+            'The left rule must be an instance of the `ContinuousRule` ': false
+            class.: false
+            'The right rule must be an instance of the `ContinuousRule` ': false
+        def `description`:
+            ∈ %s%.3f, %.3f%s: false
+            [: false
+            (: false
+            ]: false
+            ): false
+        def `__str__`:
+            %s ∈ %s%.3f, %.3f%s: false
+            [: false
+            (: false
+            ]: false
+            ): false
+        def `__repr__`:
+            IntervalRule(attr_name='%s', left_rule=%s, right_rule=%s): false
+utils/tree/skltreeadapter.py:
+    class `SklTreeAdapter`:
+        def `rules`:
+            values: false
+            _: false
 widgets/__init__.py:
     def `widget_discovery`:
         Orange3: false
@@ -15880,6 +15933,8 @@ widgets/visualize/utils/scene.py:
     class `UpdateItemsOnSelectGraphicsScene`:
         def `__handle_selection`:
             selection_changed: false
+widgets/visualize/utils/tree.py:
+    {__name__} module was moved. Use {tree.__name__} instead: false
 widgets/visualize/utils/vizrank.py:
     class `VizRankDialog`:
         Score Plots: Oceni prikaze
@@ -15979,56 +16034,3 @@ widgets/visualize/utils/widget.py:
     class `OWProjectionWidgetWithName`:
         projection: false
     iris: false
-widgets/visualize/utils/tree/rules.py:
-    class `DiscreteRule`:
-        def `merge_with`:
-            Merged two discrete rules `%s` and `%s`: false
-        def `description`:
-            {} {}: false
-            =: false
-            ≠: false
-        def `__str__`:
-            {} {} {}: false
-            =: false
-            ≠: false
-        def `__repr__`:
-            DiscreteRule(attr_name='%s', equals=%s, value=%s): false
-    class `ContinuousRule`:
-        def `merge_with`:
-            'Continuous rules can currently only be ': false
-            merged with other continuous rules: false
-        def `description`:
-            %s %.3f: false
-            >: false
-            ≤: false
-        def `__str__`:
-            %s %s %.3f: false
-            >: false
-            ≤: false
-        def `__repr__`:
-            "ContinuousRule(attr_name='%s', greater=%s, value=%s, ": false
-            inclusive=%s): false
-    class `IntervalRule`:
-        def `__init__`:
-            'The left rule must be an instance of the `ContinuousRule` ': false
-            class.: false
-            'The right rule must be an instance of the `ContinuousRule` ': false
-        def `description`:
-            ∈ %s%.3f, %.3f%s: false
-            [: false
-            (: false
-            ]: false
-            ): false
-        def `__str__`:
-            %s ∈ %s%.3f, %.3f%s: false
-            [: false
-            (: false
-            ]: false
-            ): false
-        def `__repr__`:
-            IntervalRule(attr_name='%s', left_rule=%s, right_rule=%s): false
-widgets/visualize/utils/tree/skltreeadapter.py:
-    class `SklTreeAdapter`:
-        def `rules`:
-            values: false
-            _: false
